Use Slang to unpack verilog structs in C++ (#46)
* use slang-reflect to use verilog structs in C

* use slang-reflect only if available

* install slang in docker image

+++ sim/Inst.hpp
@@ -18,7 +18,7 @@ struct Inst
     uint8_t sqn;
     uint8_t fu;
     uint8_t rd;
-    uint8_t tag;
+    uint16_t tag;
     uint8_t flags;
     uint8_t interruptCause;
     uint8_t retIdx;
